refactoring common to shared
