+++ docker/lib/dependabot/docker.rb
@@ -1,37 +1,21 @@
 # typed: strong
 # frozen_string_literal: true
 
-require "zeitwerk"
-
-loader = Zeitwerk::Loader.new
-
-# Set autoload paths for common/lib, excluding files whose content does not match the filename
-loader.push_dir(File.join(__dir__, "../../../common/lib"))
-loader.ignore(File.join(__dir__, "../../../common/lib/dependabot/errors.rb"))
-loader.ignore(File.join(__dir__, "../../../common/lib/dependabot/logger.rb"))
-loader.ignore(File.join(__dir__, "../../../common/lib/dependabot/notices.rb"))
-loader.ignore(File.join(__dir__, "../../../common/lib/dependabot/clients/codecommit.rb"))
-
-loader.push_dir(File.join(__dir__, ".."))
-loader.ignore("#{__dir__}/../script", "#{__dir__}/../spec", "#{__dir__}/../dependabot-docker.gemspec", "docker_compose")
-
-loader.on_load do |_file|
-  require "json"
-  require "sorbet-runtime"
-  require "dependabot/errors"
-  require "dependabot/logger"
-  require "dependabot/notices"
-  require "dependabot/clients/codecommit"
-end
-
-loader.log! if ENV["DEBUG"]
-loader.setup
+# These all need to be required so the various classes can be registered in a
+# lookup table of package manager names to concrete classes.
+require "dependabot/docker/file_fetcher"
+require "dependabot/docker/file_parser"
+require "dependabot/docker/update_checker"
+require "dependabot/docker/file_updater"
+require "dependabot/docker/metadata_finder"
+require "dependabot/docker/requirement"
+require "dependabot/docker/version"
 
 require_relative "docker_compose"
 
+require "dependabot/pull_request_creator/labeler"
 Dependabot::PullRequestCreator::Labeler
   .register_label_details("docker", name: "docker", colour: "21ceff")
 
+require "dependabot/dependency"
 Dependabot::Dependency.register_production_check("docker", ->(_) { true })
-
-
